Crate jetstream_error

Error handling for Jetstream

Structs

Error

Error represents a collection of errors that occurred during a Jetstream operation.

Traits

IntoError

IntoError trait allows conversion of various error types into Jetstream Error.

Type Aliases

Result

Result is a type alias for a Result type that uses the Error type.